Typecast changed to double... Unsigned int is so big ...

+++ fibonacci-lut/fibonacci-lut.cpp
@@ -48,7 +48,7 @@ void FibonacciLookupTable::calculate_fibo(unsigned int index) {
 	if (*(checker_ptr + (index-1)) == false) {
 		//Uses Moivre-Binet
 		*(first_fibo_ptr + (index-1)) = (1.0/std::sqrt(5.0) * //kein std::floor bei ungenäherter formel
-			(std::pow((0.5*(1+std::sqrt(5.0))),(long int)(index-1)) - std::pow((0.5*(1-std::sqrt(5.0))),(long int)(index-1))));
+			(std::pow((0.5*(1+std::sqrt(5.0))),(double)(index-1)) - std::pow((0.5*(1-std::sqrt(5.0))),(double)(index-1))));
 		*(checker_ptr + (index-1)) = true;
 		std::cout << "\nWert berechnet und abgespeichert.\n";
 	} else {
